Got rid of unnecessary default constructor for Enemy class.
Now classifying the captured fighter as an enemy.

Added a new path_plotter() method that returns a vector of int pairs representing locations along the path. This still needs implementation.

commented out the for loop loading in the aliens in the setup() method. For some reason this is not being liked by the compiler.