Add notes for predicted future-needed functions and variables
I currently plan on representing the enemy custom paths as a queue of pairs representing the change in x, y. With a queue, we can easily loop the change in direction for each "step" of the enemy path.

Also, a function that tells each enemy to return to the formation, this could possibly be a boolean